sql/requirements.txt
Faker
factory_boy
prettyprint